big changes
Sooo, I've been working on a WYSIWYG editor for Skal so that it'd be easier for him to see what hes changing while he's changing it.

So to make that work I'm returning the site to pure html and css (with only a little bit of js, instead of a tonne).

I took this time to add the Artibus page and get that going too. It should look exactly the same as before, if it doesn't I'll have more work to do kek